Moved interceptor to services folder
