+++ src/app/app.module.ts
@@ -11,7 +11,6 @@ import { environment } from '../environments/environment'
 import { AdminModule } from './admin/admin.module'
 import { AppRoutingModule } from './app-routing.module'
 import { AppComponent } from './app.component'
-import { HttpMockRequestInterceptor } from './auth/http-mock-request.interceptor'
 import { EmployeeModule } from './employee/employee.module'
 import { ErrorComponent } from './error/error.component'
 import { ChartComponent } from './home/chart/chart.component'
@@ -23,6 +22,7 @@ import { MessagingModule } from './messaging/messaging.module'
 import { PipeModule } from './pipes/pipe.module'
 import { AuthService } from './services/auth/auth.service'
 import { EmployeesService } from './services/employees/employees.service'
+import { HttpMockRequestInterceptor } from './services/interceptor/http-mock-request.interceptor'
 import { SkillCategoriesService } from './services/skill-categories/skill-categories.service'
 import { SkillsService } from './services/skills/skills.service'
